# Conflicts:
#	sadl3/com.ge.research.sadl.parent/com.ge.research.jena/lib/reasoner-api.jar
#	sadl3/com.ge.research.sadl.parent/com.ge.research.jena/lib/reasoner-impl.jar

+++ sadl3/com.ge.research.sadl.parent/com.ge.research.sadl.reasoner-impl/src/main/java/com/ge/research/sadl/model/visualizer/GraphVizVisualizer.java
@@ -31,6 +31,8 @@ public class GraphVizVisualizer implements IGraphVisualizer {
 	private boolean repeatObjNode;
 	private boolean repeatSubjNode;
 	private List<String> nodes;
+	private HashMap<Object, List<Object[]>> subjectList;
+	private HashMap<Object, List<Object[]>> objectList;
 
 	@Override
 	public void initialize(String tempDir, String bfn, String graphName, String anchorNode, Orientation orientation, String description) {
@@ -122,10 +124,13 @@ public class GraphVizVisualizer implements IGraphVisualizer {
 				}
 			}
 			else {
-				// delete the .dot file
-				File dotfile = new File(dotfilepath);
-				if (dotfile.exists()) {
-					dotfile.delete();
+				String dontdelete = System.getenv("GraphVizKeepDot");
+				if (dontdelete == null || !dontdelete.equalsIgnoreCase("true")) {
+					// delete the .dot file
+					File dotfile = new File(dotfilepath);
+					if (dotfile.exists()) {
+						dotfile.delete();
+					}
 				}
 			}
 		}
@@ -191,6 +196,21 @@ public class GraphVizVisualizer implements IGraphVisualizer {
 		nothingCount = 0;
 		graphedSubjectMap = null;
 		duplicateObjectMap = null;
+		
+		subjectList = new HashMap<Object, List<Object[]>>();
+		objectList = new HashMap<Object, List<Object[]>>();
+		while (rs.hasNext()) {
+			Object[] row = rs.next();
+			if (row[0] != null) {
+				addRowToMap(subjectList, rs.getShowNamespaces() ? row[0] : rs.extractLocalName(row[0]), row);
+//				subjectList.add(rs.getShowNamespaces() ? row[0] : rs.extractLocalName(row[0]));
+			}
+			if (row[2] != null) {
+				addRowToMap(objectList, rs.getShowNamespaces() ? row[2] : rs.extractLocalName(row[2]), row);
+//				objectList.add(rs.getShowNamespaces() ? row[2] : rs.extractLocalName(row[2]));
+			}
+		}
+		rs.first();	// reset cursor
 		
 		while (rs.hasNext()) {
 			Object[] row = rs.next();
@@ -278,6 +298,17 @@ public class GraphVizVisualizer implements IGraphVisualizer {
 		return dotFile;
 	}
 	
+	private void addRowToMap(HashMap<Object, List<Object[]>> list, Object key, Object[] row) {
+		if (list.containsKey(key)) {
+			list.get(key).add(row);
+		}
+		else {
+			List<Object[]> valList = new ArrayList<Object[]>();
+			valList.add(row);
+			list.put(key, valList);
+		}	
+	}
+
 	private void createGraphTriple(Object s, String edgeLbl, Object o, Map<Integer, String> headAttributes, Map<Integer, String> edgeAttributes, Map<Integer, 
 			String> tailAttributes, Object[] row, String anchorNodeLabel, String subjectNode) {
 		String slbl = subjectNode;
@@ -374,6 +405,12 @@ public class GraphVizVisualizer implements IGraphVisualizer {
 				anchored = true;
 			}
 			applyAttributesToNode(headAttributes, row, anchored);
+			if (objectList.containsKey(s)) {
+				List<Object[]> valList = objectList.get(s);
+				for (int i = 0; i < valList.size(); i++) {
+					applyAttributesToNode(tailAttributes, valList.get(i), anchored);
+				}
+			}
 			sb.append("];\n");
 		}
 		if (!repeatObjNode) {
@@ -392,6 +429,12 @@ public class GraphVizVisualizer implements IGraphVisualizer {
 				anchored = true;
 			}
 			applyAttributesToNode(tailAttributes, row, anchored);
+			if (subjectList.containsKey(o)) {
+				List<Object[]> valList = subjectList.get(o);
+				for (int i = 0; i < valList.size(); i++) {
+					applyAttributesToNode(headAttributes, valList.get(i), anchored);
+				}
+			}
 			sb.append("];\n");
 		}
 		sb.append("     ");
@@ -465,7 +508,14 @@ public class GraphVizVisualizer implements IGraphVisualizer {
 						sb.append("\"");
 					}
 					else {
-						sb.append(row[key.intValue()]);
+						Object rval = row[key.intValue()];
+						if (needsQuotes(rval)) {
+							rval = escapeQuotes(rval.toString());
+							sb.append("\"" + rval + "\"");
+						}
+						else {
+							sb.append(rval);
+						}
 					}
 				}
 			}
@@ -481,6 +531,40 @@ public class GraphVizVisualizer implements IGraphVisualizer {
 		}
 	}
 	
+	private Object escapeQuotes(String str) {
+		if (str.indexOf('\"') > 0) {
+			str = str.replace("\"", "\\\"");
+		}
+		return str;
+	}
+
+	/**
+	 * Method to detect if an attribute value needs to be quoted.
+	 * @param rval
+	 * @return
+	 */
+	private boolean needsQuotes(Object rval) {
+		if (rval instanceof String) {
+			if (rval.toString().indexOf(" ") >= 0) {
+				return true;
+			}
+			String str = (String)rval;
+			for(int i =0; i<str.length(); i++) {
+				char ch = str.charAt(i);
+				if (Character.isWhitespace(ch) || Character.isISOControl(ch)) {
+					return true;
+				}
+				if (ch == '.'){
+					return true;
+				} 
+				if (ch == '-') {
+					return true;
+				}
+			}		
+		}
+		return false;
+	}
+
 	private long getSequenceNumber(Map<Integer, String> attributes, Object[] row) {
 		if (attributes != null && attributes.containsValue("sequence_number")) {
 			Iterator<Integer> keyitr = attributes.keySet().iterator();
